Fix proxy in Jadx download, fix proxy in app func, fix frida server verify download, fix upstream proxy function
