+++ docker/docker-compose_swarm.yml
@@ -29,6 +29,12 @@ services:
       - POSTGRES_HOST=postgres
       - POSTGRES_PORT=5432
       - MOBSF_API_KEY_FILE=/run/secrets/mobsf_api_key
+      #Proxy options - use it if you behind proxy
+      #- MOBSF_UPSTREAM_PROXY_ENABLED=True  
+      #- MOBSF_UPSTREAM_PROXY_SSL_VERIFY=True
+      #- MOBSF_UPSTREAM_PROXY_TYPE=http
+      #- MOBSF_UPSTREAM_PROXY_IP=host.docker.internal
+      #- MOBSF_UPSTREAM_PROXY_PORT=8080
     healthcheck:
       test: curl -f http://localhost:8000/login/ || exit 1
     depends_on:
