+++ mobsf/DynamicAnalyzer/tools/webproxy.py
@@ -8,7 +8,7 @@ import requests
 
 from django.conf import settings
 
-from mobsf.MobSF.utils import upstream_proxy
+from mobsf.MobSF.proxy import upstream_proxy
 
 logger = logging.getLogger(__name__)
