+++ mobsf/MalwareAnalyzer/views/VirusTotal.py
@@ -4,11 +4,12 @@ import requests
 
 from django.conf import settings
 
+from mobsf.MobSF.proxy import upstream_proxy
+
 from mobsf.MobSF.utils import (
     append_scan_status,
     file_size,
     get_config_loc,
-    upstream_proxy,
 )
 
 logger = logging.getLogger(__name__)
