+++ mobsf/MobSF/settings.py
@@ -501,7 +501,7 @@ else:
     UPSTREAM_PROXY_ENABLED = bool(os.getenv(
         'MOBSF_UPSTREAM_PROXY_ENABLED', ''))
     UPSTREAM_PROXY_SSL_VERIFY = os.getenv(
-        'MOBSF_UPSTREAM_PROXY_SSL_VERIFY', '1')
+        'MOBSF_UPSTREAM_PROXY_SSL_VERIFY', 'True')
     UPSTREAM_PROXY_TYPE = os.getenv('MOBSF_UPSTREAM_PROXY_TYPE', 'http')
     UPSTREAM_PROXY_IP = os.getenv('MOBSF_UPSTREAM_PROXY_IP', '127.0.0.1')
     UPSTREAM_PROXY_PORT = int(os.getenv('MOBSF_UPSTREAM_PROXY_PORT', '3128'))
