+++ mobsf/MobSF/utils.py
@@ -35,10 +35,16 @@ import psutil
 
 import requests
 
+from mobsf.MobSF.proxy import (
+    upstream_proxy,
+    docker_translate_localhost,
+)
+
 from django.shortcuts import render
 from django.utils import timezone
 
 from mobsf.StaticAnalyzer.models import RecentScansDB
+
 from mobsf.MobSF. init import api_key
 
 from . import settings
@@ -71,31 +77,6 @@ class Color(object):
     RED = '\033[91m'
     BOLD = '\033[1m'
     END = '\033[0m'
-
-
-def upstream_proxy(flaw_type):
-    """Set upstream Proxy if needed."""
-    if settings.UPSTREAM_PROXY_ENABLED:
-        if not settings.UPSTREAM_PROXY_USERNAME:
-            proxy_port = str(settings.UPSTREAM_PROXY_PORT)
-            proxy_host = '{}://{}:{}'.format(
-                settings.UPSTREAM_PROXY_TYPE,
-                docker_translate_proxy_ip(settings.UPSTREAM_PROXY_IP),
-                proxy_port)
-            proxies = {flaw_type: proxy_host}
-        else:
-            proxy_port = str(settings.UPSTREAM_PROXY_PORT)
-            proxy_host = '{}://{}:{}@{}:{}'.format(
-                settings.UPSTREAM_PROXY_TYPE,
-                settings.UPSTREAM_PROXY_USERNAME,
-                settings.UPSTREAM_PROXY_PASSWORD,
-                docker_translate_proxy_ip(settings.UPSTREAM_PROXY_IP),
-                proxy_port)
-            proxies = {flaw_type: proxy_host}
-    else:
-        proxies = {flaw_type: None}
-    verify = settings.UPSTREAM_PROXY_SSL_VERIFY in ('1', '"1"')
-    return proxies, verify
 
 
 def get_system_resources():
@@ -377,39 +358,6 @@ def find_process_by(name):
         if (name == p.info['name']):
             proc.add(p.exe())
     return proc
-
-
-def docker_translate_localhost(identifier):
-    """Convert localhost to host.docker.internal."""
-    if not identifier:
-        return identifier
-    if not os.getenv('MOBSF_PLATFORM') == 'docker':
-        return identifier
-    try:
-        identifier = identifier.strip()
-        docker_internal = 'host.docker.internal:'
-        if re.match(r'^emulator-\d{4}$', identifier):
-            adb_port = int(identifier.split('emulator-')[1]) + 1
-            # ADB port is console port + 1
-            return f'{docker_internal}{adb_port}'
-        m = re.match(r'^(localhost|127\.0\.0\.1):\d{1,5}$', identifier)
-        if m:
-            adb_port = int(identifier.split(m.group(1))[1].replace(':', ''))
-            return f'{docker_internal}{adb_port}'
-        return identifier
-    except Exception:
-        logger.exception('Failed to convert device '
-                         'identifier for docker connectivity')
-        return identifier
-
-
-def docker_translate_proxy_ip(ip):
-    """Convert localhost proxy ip to host.docker.internal."""
-    if not os.getenv('MOBSF_PLATFORM') == 'docker':
-        return ip
-    if ip and ip.strip() in ('127.0.0.1', 'localhost'):
-        return 'host.docker.internal'
-    return ip
 
 
 def get_device():
