+++ mobsf/StaticAnalyzer/views/common/pdf.py
@@ -17,8 +17,10 @@ from mobsf.MobSF import settings
 from mobsf.MobSF.utils import (
     is_md5,
     print_n_send_error_response,
-    upstream_proxy,
 )
+
+from mobsf.MobSF.proxy import upstream_proxy
+
 from mobsf.StaticAnalyzer.models import (
     RecentScansDB,
     StaticAnalyzerAndroid,
