+++ mobsf/StaticAnalyzer/views/ios/appstore.py
@@ -7,8 +7,9 @@ from django.conf import settings
 
 from mobsf.MobSF.utils import (
     append_scan_status,
-    upstream_proxy,
 )
+
+from mobsf.MobSF.proxy import upstream_proxy
 
 logger = logging.getLogger(__name__)
