+++ nginx.conf
@@ -1,0 +1,42 @@
+user  nginx;
+events {
+    worker_connections   1000;
+}
+
+http {
+    client_max_body_size 256M;
+    upstream mobsf_upstream {
+        server mobsf:8000;
+        server mobsf:1337;
+        keepalive 16;
+    }
+
+    map $server_port $forwarded_port {
+        4000 443;
+        4001 443;
+    }
+
+    proxy_set_header  X-Forwarded-Host    $host;
+    proxy_set_header  X-Forwarded-Port    $forwarded_port;
+    proxy_set_header  X-Forwarded-For     $proxy_add_x_forwarded_for;
+    proxy_redirect    off;
+    proxy_buffering   on;
+
+    server {
+        listen 4000;
+        location / {
+            proxy_pass          http://mobsf:8000;
+            proxy_read_timeout  900;
+            client_max_body_size 256M;
+        }
+    }
+
+    server {
+        listen 4001;
+        location / {
+            proxy_pass          http://mobsf:1337;
+            proxy_read_timeout  120;
+            client_max_body_size 10M;
+        }
+    }
+}
